Small refactorings.
