+++ client/src/WebViews/UpdateNoticePanel.ts
@@ -1,5 +1,5 @@
 import * as vscode from "vscode";
-import { getNonce } from "../getNonce";
+import { getNonce } from "./getNonce";
 
 
 export class UpdateNoticePanel {
