+++ client/src/WebViews/getNonce.ts
