--- client/src/WebViews/main.js
@@ -1,1 +1,0 @@
-// Prepared for later usage
